DRY up views
